+++ app/controllers/account/login_controller.rb
@@ -2,12 +2,7 @@
 
 module Account
   class LoginController < ApplicationController
-    before_action :login_required, except: [
-      :new,
-      :create,
-      :email_new_password,
-      :new_password_request
-    ]
+    before_action :login_required, except: [:new, :create]
 
     # the login form
     def new
@@ -16,40 +11,23 @@ module Account
       else
         @login = ""
         @remember = true
-        render_new_phlex
+        render_new_view
       end
     end
 
     # login post action
     def create
-      render_new_phlex and return unless params[:user]
+      render_new_view_invalid and return unless params[:user]
 
       normalize_login_params
       user = User.authenticate(login: @login, password: @password)
 
       unless user
         flash_error(:runtime_login_failed.t)
-        render_new_phlex and return
+        render_new_view_invalid and return
       end
 
       user.verified ? login_success(user) : login_unverified(user)
-    end
-
-    def email_new_password
-      @new_user = User.new
-      render_email_new_password_phlex
-    end
-
-    def new_password_request
-      @login = params[:new_user] && params[:new_user][:login]
-      @new_user = User.where("login = ? OR name = ? OR email = ?",
-                             @login, @login, @login).first
-      if @new_user.nil?
-        flash_error(:runtime_email_new_password_failed.t(user: @login))
-        render_email_new_password_phlex and return
-      else
-        set_random_password_for_new_user_and_email_them
-      end
     end
 
     # This is used to test the autologin feature.
@@ -61,16 +39,10 @@ module Account
 
     private
 
-    def render_new_phlex
+    def render_new_view(status: :ok, **render_opts)
       render(Views::Controllers::Account::Login::New.new(
                login: @login, remember: @remember
-             ))
-    end
-
-    def render_email_new_password_phlex
-      render(Views::Controllers::Account::Login::EmailNewPassword.new(
-               new_user: @new_user
-             ))
+             ), status: status, **render_opts)
     end
 
     def normalize_login_params
@@ -95,21 +67,6 @@ module Account
       render(Views::Controllers::Account::Verifications::Reverify.new(
                unverified_user: @unverified_user
              ))
-    end
-
-    def set_random_password_for_new_user_and_email_them
-      password = String.random(10)
-      @new_user.change_password(password)
-      if @new_user.save
-        flash_notice(:runtime_email_new_password_success.tp +
-                     :email_spam_notice.tp)
-        # Migrated from QueuedEmail::Password to ActionMailer + ActiveJob.
-        # See .claude/deliver_later_migration_plan.md for details.
-        PasswordMailer.build(receiver: @new_user, password:).deliver_later
-        render_new_phlex
-      else
-        flash_object_errors(@new_user)
-      end
     end
   end
 end
